requirements.txt
pandas
numpy
pymongo
requests
python-dotenv
pymongo[srv]
